requirements.txt
# Bes2 Marketer - AI Marketing Automation
# Python 3.10+

# Frontend
streamlit==1.40.1

# Database
supabase==2.10.0

# AI
google-generativeai==0.8.3

# YouTube API
google-api-python-client==2.154.0
youtube-transcript-api==0.6.2

# Utilities
python-dotenv==1.0.1
pandas==2.2.3

